feat: split chain providers
